task comments should not be deletable, creatable or editable when the underlying process is not running anymore

+++ workflix/backend/backend/src/main/kotlin/de/se/team3/logic/container/TaskCommentsContainer.kt
@@ -3,8 +3,11 @@ package de.se.team3.logic.container
 import de.se.team3.logic.DAOInterfaces.TaskCommentsDAOInterface
 import de.se.team3.logic.domain.Task
 import de.se.team3.logic.domain.TaskComment
+import de.se.team3.logic.domain.User
+import de.se.team3.logic.exceptions.AlreadyClosedException
 import de.se.team3.logic.exceptions.NotFoundException
 import de.se.team3.persistence.daos.TaskCommentsDAO
+import de.se.team3.persistence.meta.TaskCommentsTable.taskId
 import de.se.team3.webservice.containerInterfaces.TaskCommentsContainerInterface
 
 object TaskCommentsContainer : TaskCommentsContainerInterface {
@@ -14,12 +17,15 @@ object TaskCommentsContainer : TaskCommentsContainerInterface {
     // Maps the id of a task comment to the id of the task it belongs to.
     private val taskCommentsTaskIdCache = HashMap<Int, Int>()
 
+    // Maps the id of a task comment to the id of its creator
+    private val taskCommentsCreatorIdCache = HashMap<Int, String>()
+
     /**
      * Returns the task the specified task comment belongs to.
      *
      * @throws NotFoundException Is thrown if the specified task comment does not exist.
      */
-    private fun getTaskByTaskCommentId(taskCommentId: Int): Task {
+    override fun getTaskByTaskCommentId(taskCommentId: Int): Task {
         val taskId = if (taskCommentsTaskIdCache.containsKey(taskCommentId)) {
             taskCommentsTaskIdCache.get(taskCommentId)!!
         } else {
@@ -30,16 +36,31 @@ object TaskCommentsContainer : TaskCommentsContainerInterface {
     }
 
     /**
+     * Returns the creator the specified task comment was created by.
+     *
+     * @throws NotFoundException Is thrown if the specified task comment does not exist.
+     */
+    override fun getCreatorByTaskCommentId(taskCommentId: Int): User {
+        val creatorId = if (taskCommentsCreatorIdCache.containsKey(taskCommentId)) {
+            taskCommentsCreatorIdCache.get(taskCommentId)!!
+        } else {
+            taskCommentsDAO.getCreatorIdByTaskCommentId(taskCommentId)
+                ?: throw NotFoundException("task comment does not exist")
+        }
+        return UserContainer.getUser(creatorId)
+    }
+
+    /**
      * Creates the given task comment.
      *
-     * @throws NotFoundException Is thrown if the task specified in taskComment does not exist.
-     * @throws NotFoundException Is thrown if the user specified in taskComment does not exist.
+     * @throws AlreadyClosedException Is thrown if the process the comment is addressed to
+     * is not running anymore.
      */
     override fun createTaskComment(taskComment: TaskComment): Int {
         val task = TasksContainer.getTask(taskComment.taskId!!) // throws NotFoundException
 
-        if (!UserContainer.hasUser(taskComment.creatorId!!))
-            throw NotFoundException("the creator does not exist")
+        if (!task.process!!.isRunning())
+            throw AlreadyClosedException("the process the comment is addressed to is not running")
 
         val newId = taskCommentsDAO.createTaskComment(taskComment)
         taskCommentsTaskIdCache.put(newId, taskComment.taskId!!)
@@ -53,11 +74,15 @@ object TaskCommentsContainer : TaskCommentsContainerInterface {
     /**
      * Updates the given task comment.
      *
+     * @throws AlreadyClosedException Is thrown if the process the comment is addressed to
+     * is not running anymore.
      * @throws NotFoundException Is thrown if the given task comment does not exist.
      */
     override fun updateTaskComment(taskComment: TaskComment) {
         val task = getTaskByTaskCommentId(taskComment.id!!)
 
+        if (!task.process!!.isRunning())
+            throw AlreadyClosedException("the process the comment is addressed to is not running")
         val exists = taskCommentsDAO.updateTaskComment(taskComment)
         if (!exists)
             throw NotFoundException("task comment not found")
@@ -69,11 +94,15 @@ object TaskCommentsContainer : TaskCommentsContainerInterface {
     /**
      * Deletes the specified task comment.
      *
+     * @throws AlreadyClosedException Is thrown if the process the comment is addressed to
+     * is not running anymore.
      * @throws NotFoundException Is thrown if the specified task comment does not exist.
      */
     override fun deleteTaskComment(taskCommentId: Int) {
         val task = getTaskByTaskCommentId(taskCommentId)
 
+        if (!task.process!!.isRunning())
+            throw AlreadyClosedException("the process the comment is addressed to is not running")
         val existed = taskCommentsDAO.deleteTaskComment(taskCommentId)
         if (!existed)
             throw NotFoundException("task comment not found")
